requirements.txt
langchain==0.1.14
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
GitPython==3.1.42
sentence-transformers==2.3.1
chromadb==0.4.22
